move titleCase to utils
